requirements.txt
flask
streamlit
kafka-python
numpy
pandas
scikit-learn
tensorflow
psycopg2-binary
plotly
